+++ sleek_storage/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.1.0
+* BREAKING: `Box.watch` now emits `null` when value is deleted or when box is cleared.
 * BREAKING: `fromJson` now also provides the key.
 * New `Box.containsKey` method.
+* Fix `Box.putAll` method.
 
 ## 1.0.0
 * First stable release.